Working video objects w/ statistics
